[FLINK-4562] [table] Move table examples into a dedicated module in flink-examples.
This closes #2460.
