+++ flink-examples/flink-examples-table/src/main/java/org/apache/flink/table/examples/java/WordCountTable.java
@@ -17,11 +17,11 @@
  */
 package org.apache.flink.table.examples.java;
 
-import org.apache.flink.table.api.Table;
 import org.apache.flink.api.java.DataSet;
 import org.apache.flink.api.java.ExecutionEnvironment;
-import org.apache.flink.table.api.java.BatchTableEnvironment;
+import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableEnvironment;
+import org.apache.flink.table.api.java.BatchTableEnvironment;
 
 /**
   * Simple example for demonstrating the use of the Table API for a Word Count in Java.
